Added automagic blacklisting (for m minutes after n failed attempts) to the SSH frontend, by request from bl (and most of the Usual Gang of Idiots). Seems to work.
Room for improvement. Didn't want to tie the SSH server too tight to anything else, but it'd be possible to build a few commands to handle the list. Also, pulling m and n from server settings might be a good idea, next time I'm in the mood to tinker with this.